% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{assign_instance}
\alias{assign_instance}
\title{Assign a registered instance to a layer}
\usage{
assign_instance(InstanceId, LayerIds)
}
\arguments{
\item{InstanceId}{The instance ID.}

\item{LayerIds}{The layer ID, which must correspond to a custom layer. You cannot assign a registered instance to a built-in layer.}
}
\description{
Assign a registered instance to a layer.
}
\details{
\itemize{
\item You can assign registered on-premises instances to any layer type.
\item You can assign registered Amazon EC2 instances only to custom layers.
\item You cannot use this action with instances that were created with AWS OpsWorks Stacks.
}

\strong{Required Permissions}: To use this action, an AWS Identity and Access Management (IAM) user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
