% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{create_app}
\alias{create_app}
\title{Creates an app for a specified stack}
\usage{
create_app(StackId, Name, Type, Shortname = NULL, Description = NULL,
  DataSources = NULL, AppSource = NULL, Domains = NULL,
  EnableSsl = NULL, SslConfiguration = NULL, Attributes = NULL,
  Environment = NULL)
}
\arguments{
\item{StackId}{The stack ID.}

\item{Name}{The app name.}

\item{Type}{The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks Stacks deploys an application to those instances that are members of the corresponding layer. If your app isn't one of the standard types, or you prefer to implement your own Deploy recipes, specify \code{other}.}

\item{Shortname}{The app's short name.}

\item{Description}{A description of the app.}

\item{DataSources}{The app's data source.}

\item{AppSource}{A \code{Source} object that specifies the app repository.}

\item{Domains}{The app virtual host settings, with multiple domains separated by commas. For example: \code{'www.example.com, example.com'}}

\item{EnableSsl}{Whether to enable SSL for the app.}

\item{SslConfiguration}{An \code{SslConfiguration} object with the SSL configuration.}

\item{Attributes}{One or more user-defined key/value pairs to be added to the stack attributes.}

\item{Environment}{An array of \code{EnvironmentVariable} objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instance. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/workingapps-creating.html#workingapps-creating-environment}{Environment Variables}.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 10 KB (10240 Bytes). This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 10KB)."

This parameter is supported only by Chef 11.10 stacks. If you have specified one or more environment variables, you cannot modify the stack's Chef version.}
}
\description{
Creates an app for a specified stack. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/workingapps-creating.html}{Creating Apps}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
