% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{delete_app}
\alias{delete_app}
\title{Deletes a specified app}
\usage{
delete_app(
  AppId = "string"
)
}
\arguments{
\item{AppId}{[required] The app ID.}
}
\description{
Deletes a specified app.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
