% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{delete_instance}
\alias{delete_instance}
\title{Deletes a specified instance, which terminates the associated Amazon EC2 instance}
\usage{
delete_instance(InstanceId, DeleteElasticIp = NULL, DeleteVolumes = NULL)
}
\arguments{
\item{InstanceId}{The instance ID.}

\item{DeleteElasticIp}{Whether to delete the instance Elastic IP address.}

\item{DeleteVolumes}{Whether to delete the instance's Amazon EBS volumes.}
}
\description{
Deletes a specified instance, which terminates the associated Amazon EC2 instance. You must stop an instance before you can delete it.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances-delete.html}{Deleting Instances}.

\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
