% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{deregister_ecs_cluster}
\alias{deregister_ecs_cluster}
\title{Deregisters a specified Amazon ECS cluster from a stack}
\usage{
deregister_ecs_cluster(EcsClusterArn)
}
\arguments{
\item{EcsClusterArn}{The cluster's Amazon Resource Number (ARN).}
}
\description{
Deregisters a specified Amazon ECS cluster from a stack. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/workinglayers-ecscluster.html#workinglayers-ecscluster-delete}{Resource Management}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see \url{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}.
}
\examples{

}
