% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{deregister_elastic_ip}
\alias{deregister_elastic_ip}
\title{Deregisters a specified Elastic IP address}
\usage{
deregister_elastic_ip(
  ElasticIp = "string"
)
}
\arguments{
\item{ElasticIp}{[required] The Elastic IP address.}
}
\description{
Deregisters a specified Elastic IP address. The address can then be registered by another stack. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
