% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{deregister_instance}
\alias{deregister_instance}
\title{Deregister a registered Amazon EC2 or on-premises instance}
\usage{
deregister_instance(InstanceId)
}
\arguments{
\item{InstanceId}{The instance ID.}
}
\description{
Deregister a registered Amazon EC2 or on-premises instance. This action removes the instance from the stack and returns it to your control. This action cannot be used with instances that were created with AWS OpsWorks Stacks.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
