% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{describe_agent_versions}
\alias{describe_agent_versions}
\title{Describes the available AWS OpsWorks Stacks agent versions}
\usage{
describe_agent_versions(StackId = NULL, ConfigurationManager = NULL)
}
\arguments{
\item{StackId}{The stack ID.}

\item{ConfigurationManager}{The configuration manager.}
}
\description{
Describes the available AWS OpsWorks Stacks agent versions. You must specify a stack ID or a configuration manager. \code{DescribeAgentVersions} returns a list of available agent versions for the specified stack or configuration manager.
}
\examples{

}
