% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{describe_deployments}
\alias{describe_deployments}
\title{Requests a description of a specified set of deployments}
\usage{
describe_deployments(
  StackId = "string",
  AppId = "string",
  DeploymentIds = list(
    "string"
  )
)
}
\arguments{
\item{StackId}{The stack ID. If you include this parameter, the command returns a description of the commands associated with the specified stack.}

\item{AppId}{The app ID. If you include this parameter, the command returns a description of the commands associated with the specified app.}

\item{DeploymentIds}{An array of deployment IDs to be described. If you include this parameter, the command returns a description of the specified deployments. Otherwise, it returns a description of every deployment.}
}
\description{
Requests a description of a specified set of deployments.
}
\details{
This call accepts only one resource-identifying parameter.

\strong{Required Permissions}: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
