% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{describe_elastic_ips}
\alias{describe_elastic_ips}
\title{Describes \link{Elastic IP addresses}(http://docs}
\usage{
describe_elastic_ips(
  InstanceId = "string",
  StackId = "string",
  Ips = list(
    "string"
  )
)
}
\arguments{
\item{InstanceId}{The instance ID. If you include this parameter, \code{DescribeElasticIps} returns a description of the Elastic IP addresses associated with the specified instance.}

\item{StackId}{A stack ID. If you include this parameter, \code{DescribeElasticIps} returns a description of the Elastic IP addresses that are registered with the specified stack.}

\item{Ips}{An array of Elastic IP addresses to be described. If you include this parameter, \code{DescribeElasticIps} returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.}
}
\description{
Describes \href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/elastic-ip-addresses-eip.html}{Elastic IP addresses}.
}
\details{
This call accepts only one resource-identifying parameter.

\strong{Required Permissions}: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
