% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{describe_my_user_profile}
\alias{describe_my_user_profile}
\title{Describes a user's SSH information}
\usage{
describe_my_user_profile()
}
\description{
Describes a user's SSH information.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information about user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
