% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{describe_operating_systems}
\alias{describe_operating_systems}
\title{Describes the operating systems that are supported by AWS OpsWorks Stacks}
\usage{
describe_operating_systems()
}
\description{
Describes the operating systems that are supported by AWS OpsWorks Stacks.
}
