% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{describe_rds_db_instances}
\alias{describe_rds_db_instances}
\title{Describes Amazon RDS instances}
\usage{
describe_rds_db_instances(StackId, RdsDbInstanceArns = NULL)
}
\arguments{
\item{StackId}{The ID of the stack with which the instances are registered. The operation returns descriptions of all registered Amazon RDS instances.}

\item{RdsDbInstanceArns}{An array containing the ARNs of the instances to be described.}
}
\description{
Describes Amazon RDS instances.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.

This call accepts only one resource-identifying parameter.
}
\examples{

}
