% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{describe_user_profiles}
\alias{describe_user_profiles}
\title{Describe specified users}
\usage{
describe_user_profiles(IamUserArns = NULL)
}
\arguments{
\item{IamUserArns}{An array of IAM or federated user ARNs that identify the users to be described.}
}
\description{
Describe specified users.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information about user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
