% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{get_hostname_suggestion}
\alias{get_hostname_suggestion}
\title{Gets a generated host name for the specified layer, based on the current host name theme}
\usage{
get_hostname_suggestion(
  LayerId = "string"
)
}
\arguments{
\item{LayerId}{[required] The layer ID.}
}
\description{
Gets a generated host name for the specified layer, based on the current host name theme.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
