% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_package.R
\docType{package}
\name{paws.opsworks-package}
\alias{paws.opsworks}
\alias{paws.opsworks-package}
\title{paws.opsworks: AWS OpsWorks}
\description{
Welcome to the *AWS OpsWorks Stacks API Reference*.
This guide provides descriptions, syntax, and usage examples for AWS
OpsWorks Stacks actions and data types, including common parameters
and error codes.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
