% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{reboot_instance}
\alias{reboot_instance}
\title{Reboots a specified instance}
\usage{
reboot_instance(InstanceId)
}
\arguments{
\item{InstanceId}{The instance ID.}
}
\description{
Reboots a specified instance. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances-starting.html}{Starting, Stopping, and Rebooting Instances}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
