% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{register_ecs_cluster}
\alias{register_ecs_cluster}
\title{Registers a specified Amazon ECS cluster with a stack}
\usage{
register_ecs_cluster(EcsClusterArn, StackId)
}
\arguments{
\item{EcsClusterArn}{The cluster's ARN.}

\item{StackId}{The stack ID.}
}
\description{
Registers a specified Amazon ECS cluster with a stack. You can register only one cluster with a stack. A cluster can be registered with only one stack. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/workinglayers-ecscluster.html}{Resource Management}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
