% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{set_time_based_auto_scaling}
\alias{set_time_based_auto_scaling}
\title{Specify the time-based auto scaling configuration for a specified instance}
\usage{
set_time_based_auto_scaling(
  InstanceId = "string",
  AutoScalingSchedule = list(
    Monday = list(
      "string"
    ),
    Tuesday = list(
      "string"
    ),
    Wednesday = list(
      "string"
    ),
    Thursday = list(
      "string"
    ),
    Friday = list(
      "string"
    ),
    Saturday = list(
      "string"
    ),
    Sunday = list(
      "string"
    )
  )
)
}
\arguments{
\item{InstanceId}{[required] The instance ID.}

\item{AutoScalingSchedule}{An \code{AutoScalingSchedule} with the instance schedule.}
}
\description{
Specify the time-based auto scaling configuration for a specified instance. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/workinginstances-autoscaling.html}{Managing Load with Time-based and Load-based Instances}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
