% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes tags from a specified stack or layer}
\usage{
untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{The stack or layer's Amazon Resource Number (ARN).}

\item{TagKeys}{A list of the keys of tags to be removed from a stack or layer.}
}
\description{
Removes tags from a specified stack or layer.
}
\examples{

}
