% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{update_elastic_ip}
\alias{update_elastic_ip}
\title{Updates a registered Elastic IP address's name}
\usage{
update_elastic_ip(
  ElasticIp = "string",
  Name = "string"
)
}
\arguments{
\item{ElasticIp}{[required] The IP address for which you want to update the name.}

\item{Name}{The new name.}
}
\description{
Updates a registered Elastic IP address's name. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
