% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{update_my_user_profile}
\alias{update_my_user_profile}
\title{Updates a user's SSH public key}
\usage{
update_my_user_profile(SshPublicKey = NULL)
}
\arguments{
\item{SshPublicKey}{The user's SSH public key.}
}
\description{
Updates a user's SSH public key.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have self-management enabled or an attached policy that explicitly grants permissions. For more information about user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
