% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{update_rds_db_instance}
\alias{update_rds_db_instance}
\title{Updates an Amazon RDS instance}
\usage{
update_rds_db_instance(RdsDbInstanceArn, DbUser = NULL, DbPassword = NULL)
}
\arguments{
\item{RdsDbInstanceArn}{The Amazon RDS instance's ARN.}

\item{DbUser}{The master user name.}

\item{DbPassword}{The database password.}
}
\description{
Updates an Amazon RDS instance.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
