% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.opsworks_operations.R
\name{update_volume}
\alias{update_volume}
\title{Updates an Amazon EBS volume's name or mount point}
\usage{
update_volume(
  VolumeId = "string",
  Name = "string",
  MountPoint = "string"
)
}
\arguments{
\item{VolumeId}{[required] The volume ID.}

\item{Name}{The new name.}

\item{MountPoint}{The new mount point.}
}
\description{
Updates an Amazon EBS volume's name or mount point. For more information, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.
}
\details{
\strong{Required Permissions}: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see \href{http://docs.aws.amazon.com/opsworks/latest/userguide/opsworks-security-users.html}{Managing User Permissions}.
}
\examples{

}
